refactor arrange
